# DOC/swagger-api-docs.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "Videly REST API",
        "description": "REST API сервер для мобильного приложения Videly",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://localhost:8443",
            "description": "Generated server url"
        }
    ],
...
    "paths": {
...
        "/api/subscriptions/{userId}": {
            "get": {
                "tags": [
                    "SubscriptionsController"
                ],
                "summary": "Загрузить подписки пользователя",
                "operationId": "getUserSubscriptions",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "ID пользователя",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Данные получены",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/SubscriptionDTO"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ничего не найдено",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MessageResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MessageResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "SubscriptionsController"
                ],
                "summary": "Подписаться на канал",
                "operationId": "subscribe",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "ID пользователя",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "channelId",
                        "in": "query",
                        "description": "ID канала",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MessageResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ничего не найдено",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MessageResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MessageResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/history/history/{userId}": {
            "get": {
                "tags": [
                    "HistoryController"
                ],
                "summary": "Загрузка истории просмотров",
                "description": "Загрузка истории просмотров конкреного пользоавтеля",
                "operationId": "getViewsHistory",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "ID пользователя",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Данные загружены",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/HistoryDTO"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ничего не найдено",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MessageResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MessageResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "HistoryController"
                ],
                "summary": "Добавить новое видео в историю просмотров",
                "operationId": "addToHistory",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "ID пользователя",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "string",
                                "description": "ID видео"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HistoryDTO"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Пользователь не найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MessageResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MessageResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "CommentDTO": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "content": {
                        "type": "string"
                    },
                    "writtenOn": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "deleted": {
                        "type": "boolean"
                    }
                }
            },
            "VideoDTO": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    },
                    "previewURL": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "views": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "length": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "allowComments": {
                        "type": "boolean"
                    },
                    "allowRating": {
                        "type": "boolean"
                    },
                    "videoURL": {
                        "type": "string"
                    },
                    "comments": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CommentDTO"
                        }
                    }
                }
            },
            "MessageResponse": {
                "type": "object",
                "properties": {
                    "messages": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
...
            "HistoryDTO": {
                "type": "object",
                "properties": {
                    "videos": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/VideoDTO"
                        }
                    },
                    "ts": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
            "ChannelDTO": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "avatarURL": {
                        "type": "string"
                    },
                    "headerURL": {
                        "type": "string"
                    },
                    "about": {
                        "type": "string"
                    },
                    "createdAt": {
                        "$ref": "#/components/schemas/Timestamp"
                    },
                    "videos": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/VideoDTO"
                        }
                    }
                }
            },
            "Timestamp": {
                "type": "object",
                "properties": {
                    "seconds": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "nanos": {
                        "type": "integer",
                        "format": "int32"
                    }
                }
            },
...
            "SubscriptionDTO": {
                "type": "object",
                "properties": {
                    "channels": {
                        "type": "array",
                        "description": "Каналы на которые подписан пользователь",
                        "items": {
                            "$ref": "#/components/schemas/ChannelDTO"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}